Fixes root certificate parsing errors
I was trying to avoid forcing the user to send the length of the
certificate buffer, but that turns out to be `0xabad1dea`.

The certificate errors were because, inside the library, after the
buffer had been cast to a pointer, `sizeof` just returned the size of a
pointer, usually 4 or 8 bytes, instead of the length of the data
pointed to.

Additionally, to keep resource usage to a minimum, users may want to
encode their certificate in DER instead of PEM. The mbed TLS libary
allows this, but my technique would have obscured that possibility.
To allow binary data, the user must pass the length herself.